+++ testing/test_recwarn.py
@@ -112,10 +112,9 @@ class TestDeprecatedCall(object):
         pytest.deprecated_call(self.dep_explicit, 0)
 
     def test_deprecated_call_as_context_manager_no_warning(self):
-        with pytest.raises(pytest.fail.Exception) as ex:
+        with pytest.raises(pytest.fail.Exception, matches='^DID NOT WARN'):
             with pytest.deprecated_call():
                 self.dep(1)
-        assert str(ex.value).startswith("DID NOT WARN")
 
     def test_deprecated_call_as_context_manager(self):
         with pytest.deprecated_call():